+++ ssh_proxy_server/__entrypoints__.py
@@ -19,6 +19,14 @@ entry_points = {
         'replace_file = ssh_proxy_server.plugins.sftp.replace_file:SFTPProxyReplaceHandler',
         'store_file = ssh_proxy_server.plugins.sftp.store_file:SFTPHandlerStoragePlugin'
     ],
+    'ServerTunnelBaseForwarder': [
+        'base = ssh_proxy_server.forwarders.tunnel:ServerTunnelForwarder',
+        'inject = ssh_proxy_server.plugins.tunnel.injectservertunnel:InjectableServerTunnelForwarder'
+    ],
+    'ClientTunnelBaseForwarder': [
+        'base = ssh_proxy_server.forwarders.tunnel:ClientTunnelForwarder',
+        'inject = ssh_proxy_server.plugins.tunnel.injectclienttunnel:InjectableClientTunnelForwarder'
+    ],
     'BaseServerInterface': [
         'base = ssh_proxy_server.interfaces.server:ServerInterface'
     ],