+++ ssh_proxy_server/cli.py
@@ -31,6 +31,12 @@ from ssh_proxy_server.interfaces.sftp import (
     SFTPProxyServerInterface
 )
 
+from ssh_proxy_server.forwarders.tunnel import (
+    ServerTunnelBaseForwarder,
+    ClientTunnelForwarder,
+    ServerTunnelForwarder,
+    ClientTunnelBaseForwarder
+)
 
 from ssh_proxy_server.workarounds import dropbear
 from ssh_proxy_server.plugins.ssh.mirrorshell import SSHMirrorForwarder
@@ -100,6 +106,20 @@ def main():
         default=SFTPHandlerPlugin,
         help='SFTP Handler to handle sftp file transfers',
         baseclass=SFTPHandlerBasePlugin
+    )
+    parser.add_module(
+        '--server-tunnel',
+        dest='server_tunnel_interface',
+        default=ServerTunnelForwarder,
+        help='interface to handle tunnels from the server',
+        baseclass=ServerTunnelBaseForwarder
+    )
+    parser.add_module(
+        '--client-tunnel',
+        dest='client_tunnel_interface',
+        default=ClientTunnelForwarder,
+        help='interface to handle tunnels from the client',
+        baseclass=ClientTunnelBaseForwarder
     )
     parser.add_module(
         '--auth-interface',
@@ -172,6 +192,8 @@ def main():
         scp_interface=args.scp_interface,
         sftp_interface=args.sftp_interface,
         sftp_handler=args.sftp_handler,
+        server_tunnel_interface=args.server_tunnel_interface,
+        client_tunnel_interface=args.client_tunnel_interface,
         authentication_interface=args.auth_interface,
         authenticator=args.authenticator,
         transparent=args.transparent,