+++ ssh_proxy_server/server.py
@@ -30,6 +30,8 @@ class SSHProxyServer:
         scp_interface=None,
         sftp_interface=None,
         sftp_handler=None,
+        server_tunnel_interface=None,
+        client_tunnel_interface=None,
         authentication_interface=None,
         authenticator=None,
         transparent=False,
@@ -53,6 +55,8 @@ class SSHProxyServer:
         self.scp_interface = scp_interface
         self.sftp_handler = sftp_handler
         self.sftp_interface = self.sftp_handler.get_interface() or sftp_interface
+        self.server_tunnel_interface = server_tunnel_interface
+        self.client_tunnel_interface = client_tunnel_interface
         # Server Interface
         self.authentication_interface = authentication_interface
         self.authenticator = authenticator